requirements.txt
Django==5.1.5
gunicorn==23.0.0
whitenoise== 6.9.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
Pillow==11.1.0
django-environ==0.12.0
django-cors-headers==4.7.0

